+++ Core/Entities/Characters/enemy/enemyBehavior/ChaseMotor.cs
@@ -11,7 +11,7 @@ public class ChaseMotor : Motor
         TargetPosition = targetPosition;
     }
 
-    public override Vector2 Move(Vector2 position, int velocity)
+    public override Vector2 Move(Vector2 position, float velocity)
     {
         Vector2 vec = Vector2.Zero;
